+++ src/candlewick/multibody/Multibody.h
@@ -26,5 +26,22 @@ namespace multibody {
   void guiAddPinocchioModelInfo(entt::registry &reg, const pin::Model &model,
                                 const pin::GeometryModel &geom_model,
                                 int table_height_lines = 6);
+
+  struct PinGeomObjComponent {
+    pin::GeomIndex geom_index;
+    operator auto() const { return geom_index; }
+  };
+
+  struct PinFrameComponent {
+    pin::FrameIndex frame_id;
+    operator auto() const { return frame_id; }
+  };
+
+  /// Similar to PinFrameComponent, but tags the entity to render the frame
+  /// velocity - not its placement as e.g. a triad.
+  struct PinFrameVelocityComponent {
+    pin::FrameIndex frame_id;
+    operator auto() const { return frame_id; }
+  };
 } // namespace multibody
 } // namespace candlewick
